requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
asyncpg==0.18.3
attrs==19.1.0
chardet==3.0.4
idna==2.8
multidict==4.5.2
psycopg2-binary==2.8.3
yarl==1.3.0
